Rewrote protocol handler to use the new chunk  deserializer.
Protocol handler is now a `gen_server` that is fully reactive to BEAM messages.

Chunk deserializer had to be modified to no longer take an MFA in, because there's no way to make sure the MFA gets called with the latest protocol handler state without using a beam message, which is unnecessary overhead.